\name{KalmanRcppDemo-package}
\alias{KalmanRcppDemo-package}
\alias{KalmanRcppDemo}
\docType{package}
\title{
  A simple implementation of Kalman filter 
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Maintainer: Amanda <amandajayanetti@gmail.com>
}
\references{
  For details on Kalman filter: http://www.bzarg.com/p/how-a-kalman-filter-works-in-pictures/
}